refactor: split format/cos.go view-model build from render (#4657)
format/cos.go had grown to 865 LOC with FormatCoSInterfaceSummary
interleaving the config+runtime view-model join with every render
block (per-interface header, binding telemetry, the tabwriter queue
grid, and the interleaved per-queue Drops / Sojourn / Surplus /
Equal-flow / OwnerProfile / DrainShape / Waterfill detail rows). This
mirrors the #4656 status.go aggregate-then-render split.

Pure code motion, output byte-identical. The change:

  - Move the view-model construction into a new cos_sections.go:
    configuredCoSInterfaceViews, cosRuntimeIndex + buildCoSRuntimeIndex
    + lookup + cosRuntimeCandidateNames, buildCoSQueueViews, and
    isOldJSONSyntheticDefaultQueue are the only places that join the
    CoS/interface stanzas with the runtime snapshot.

  - Decompose the FormatCoSInterfaceSummary render loop into per-section
    helpers (also in cos_sections.go): writeCoSInterfaceHeader (config +
    runtime + waterfill header), writeCoSQueueTable (the aligned
    tabwriter grid), and writeCoSQueueDetail (the interleaved detail
    rows). renderBindingScopedTelemetry and the histogram-percentile
    helpers move alongside them.

  - cos.go keeps the cosInterfaceView / cosQueueView types, the
    FormatCoSInterfaceSummary orchestrator, and the leaf formatters
    (cos.go 280 LOC, cos_sections.go 632 LOC).

The queue view slice is still built exactly once per interface by the
orchestrator and shared between the binding-scoped telemetry line and
the queue table — no extra fetches, no re-iteration.

Validation: added TestFormatCoSInterfaceSummaryGolden
(testdata/cos_summary.golden) as the byte-identity gate. The golden was
generated from the pre-refactor code with a fixture that exercises every
conditional header/queue/detail block, and the refactored code
reproduces it byte-for-byte; it reuses the package-level -update flag.
go build ./... clean, go vet clean, and go test
./pkg/dataplane/userspace/format/ green (golden + all 24 existing CoS
tests). No operator-facing contract change (output unchanged), so no
module doc update is needed.
